# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Crepe & Waffle API documentation'
  description: 'A documented API for the required requests.'
  contact:
    email: ahmed.bltagy@appssquare.com
  version: 1.0.0
servers:
  -
    url: 'https://app.crepe-waffle.com/api'
    description: 'Production Server'
paths:
...
  /homeSlider:
    get:
      tags:
        - 'Sliders&gallery'
      summary: 'Retrieve home sliders for website and app'
      description: ''
      operationId: homeSlider
      responses:
        200:
          description: ''
        422:
          description: 'Validation Error'
      security:
        -
          petstore_auth:
            - 'write:pets'
            - 'read:pets'
...
  /serviceRate:
    post:
      tags:
        - Orders
      summary: 'Submit a rate for a service'
      description: 'Authentication required on header using token'
      operationId: serviceRate
      parameters:
        -
          name: user_id
          in: query
          description: 'User ID ( already retrieved )'
          required: true
          schema:
            type: integer
        -
          name: service_rate
          in: query
          description: "the object that have the rate it should look like\n\n    {'crepe':5,'waffle ':5,'delivery':5} "
          required: true
          schema:
            type: string
        -
          name: order_id
          in: query
          description: 'The Order id that you got from submitting order'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: ''
        422:
          description: 'Validation Error'
      security:
        -
          petstore_auth:
            - 'write:pets'
            - 'read:pets'
...
  /products:
    get:
      tags:
        - Products
      summary: 'Get and store product images'
      operationId: ListProducts
      parameters:
        -
          name: category_id
          in: query
          description: 'Category ID'
          required: false
          schema:
            type: integer
        -
          name: user_id
          in: query
          description: 'The user ID if available to return if the user has rated the product or not'
          required: false
          schema:
            type: integer
      responses:
        200:
          description: "\n     *     "
...
  '/products/{id}':
    get:
      tags:
        - Products
      summary: 'Show one'
      parameters:
        -
          name: id
          in: path
          description: 'Product ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: "\n     *     "
...
components:
  securitySchemes:
    'Password Based':
      type: oauth2
      description: 'Use a global client_id / client_secret and your username / password combo to obtain a token'
      in: header
      flows:
        password:
          authorizationUrl: /api/login
          tokenUrl: /api/login
          scopes: {  }
      scheme: https
...